api/requirements.txt
azure-functions
requests
microsoft-graph-for-python